flask_app/requirements.txt
Flask
python-dotenv
pytest
flake8
pylint
black
